add authorization for actions

+++ routes/campgrounds.js
@@ -3,19 +3,7 @@ const router = express.Router();
 
 const catchAsync = require('../utils/catchAsync');
 const Campground = require('../models/campground');
-const {campgroundSchema} = require('../schema');
-const {isLoggedIn} = require('../middleware');
-
-const validateCampground = (req, res, next) => {
-	const {error} = campgroundSchema.validate(req.body);
-	if (error) {
-		const msg = error.details.map(el => el.message).join(',');
-		throw new ExpressError(msg, 400);
-	} else {
-		next();
-	}
-	// console.log(result);
-};
+const {isLoggedIn, isAuthor, validateCampground} = require('../middleware');
 
 // routing
 router.get(
@@ -36,6 +24,7 @@ router.post(
 	validateCampground,
 	catchAsync(async (req, res, next) => {
 		const campground = new Campground(req.body.campground);
+		campground.author = req.user._id;
 		await campground.save();
 		req.flash('success', 'Successfully made a new campground!');
 		res.redirect(`/campgrounds/${campground._id}`);
@@ -46,7 +35,9 @@ router.get(
 	'/:id',
 	catchAsync(async (req, res, next) => {
 		const {id} = req.params;
-		const campground = await Campground.findById(id).populate('reviews');
+		const campground = await Campground.findById(id)
+			.populate({path: 'reviews', populate: {path: 'author'}})
+			.populate('author');
 		if (!campground) {
 			req.flash('error', 'Cannot find that campground');
 			res.redirect('/campgrounds');
@@ -59,9 +50,14 @@ router.get(
 router.get(
 	'/:id/edit',
 	isLoggedIn,
+	isAuthor,
 	catchAsync(async (req, res, next) => {
 		const {id} = req.params;
 		const campground = await Campground.findById(id);
+		if (!campground) {
+			req.flash('error', 'Cannot find the campground');
+			return res.redirect(`/campgrounds`);
+		}
 		res.render('campgrounds/edit', {campground});
 	})
 );
@@ -69,6 +65,7 @@ router.get(
 router.put(
 	'/:id',
 	isLoggedIn,
+	isAuthor,
 	validateCampground,
 	catchAsync(async (req, res, next) => {
 		const {id} = req.params;
@@ -83,8 +80,14 @@ router.put(
 router.delete(
 	'/:id',
 	isLoggedIn,
+	isAuthor,
 	catchAsync(async (req, res, next) => {
 		const {id} = req.params;
+		const campground = await Campground.findById(id);
+		if (!campground.author.equals(req.user._id)) {
+			req.flash('error', 'You do not have permission to do that');
+			return res.redirect(`/campgrounds/${campground._id}`);
+		}
 		await Campground.findByIdAndDelete(id);
 		req.flash('error', 'Deleted the campground!');
 		res.redirect('/campgrounds');
